Fix some bugs and make little renaming for better understanding

+++ app/src/main/java/it/mirea/ecoctrl/views/fragments/AdminFragment.java
@@ -4,24 +4,27 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.ItemTouchHelper;
+import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-
-import com.google.android.material.snackbar.Snackbar;
 
 import java.util.List;
 
-import it.mirea.ecoctrl.R;
 import it.mirea.ecoctrl.databinding.AdminFragmentBinding;
 import it.mirea.ecoctrl.di.AppExecutors;
+import it.mirea.ecoctrl.di.ServiceLocator;
+import it.mirea.ecoctrl.domain.models.Place;
 import it.mirea.ecoctrl.domain.models.User;
 import it.mirea.ecoctrl.repositories.room.MapRoomDatabase;
 import it.mirea.ecoctrl.viewModels.AdminViewModel;
@@ -36,12 +39,9 @@ public class AdminFragment extends Fragment {
     private static final String EMAIL = "email";
     private static final String LVL = "lvl";
 
-    private MapRoomDatabase mapRoomDatabase;
     private UserListAdapter listAdapter;
-    RecyclerView all_usrs;
-    List<User> users;
+    List<User> userss;
 
-    Context context;
     int back_tap;
 
     private AdminViewModel adminViewModel;
@@ -50,9 +50,9 @@ public class AdminFragment extends Fragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        mapRoomDatabase = MapRoomDatabase.getInstance(getActivity().getApplicationContext());
+
         initAdminViewModel();
-        context= new UsrActivity();
+        ServiceLocator.getInstance().initBase(getActivity().getApplication());
 
     }
 
@@ -60,6 +60,12 @@ public class AdminFragment extends Fragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         binding= AdminFragmentBinding.inflate(inflater, container, false);
+        binding.allUsers.setLayoutManager(new LinearLayoutManager(getContext()));
+        return binding.getRoot();
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
 
         binding.backButt.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -71,14 +77,12 @@ public class AdminFragment extends Fragment {
                 else
                 {
                     back_tap++;
-                    Snackbar.make(getActivity().findViewById(R.id.content), "If you want log out from your account, tap again",
-                            Snackbar.LENGTH_LONG).show();
+                    Toast.makeText(getContext(), "Кликните еще раз, если хотите выйти из аккаунта", Toast.LENGTH_SHORT).show();
                 }
             }
         });
-
-        listAdapter= new UserListAdapter(users,((UsrActivity) requireActivity()));
-        all_usrs.setAdapter(listAdapter);
+        listAdapter= new UserListAdapter(userss,((UsrActivity) requireActivity()));
+        binding.allUsers.setAdapter(listAdapter);
         new ItemTouchHelper(new ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT | ItemTouchHelper.RIGHT) {
             @Override
             public boolean onMove(RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, RecyclerView.ViewHolder target) {
@@ -97,38 +101,35 @@ public class AdminFragment extends Fragment {
 //                });
 //            }
 
-            @Override
-            //MAKE SWIPE CHECKER
-            public void onSwiped(final RecyclerView.ViewHolder viewHolder, int swipeDir) {
+            @Override            public void onSwiped(final RecyclerView.ViewHolder viewHolder, int swipeDir) {
                 AppExecutors.getInstance().diskIO().execute(new Runnable() {
                     @Override
                     public void run() {
                         int position = viewHolder.getAdapterPosition();
                         List<User> tasks = listAdapter.getData();
-                        adminViewModel.deleteUser(position,tasks);
-
+                        if(!tasks.get(position).getRole().toString().equals("Admin")){
+                            adminViewModel.deleteUser(position,tasks);
+                        }
                     }
                 });
             }
-        }).attachToRecyclerView(all_usrs);
+        }).attachToRecyclerView(binding.allUsers);
         retrievePlaces();
-        return binding.getRoot();
+
     }
 
-    private void initPlistViewModel() {
-        adminViewModel = new ViewModelProvider(this)
-                .get(AdminViewModel.class);
-    }
 
     private void retrievePlaces() {
         adminViewModel.getAllUsers();
-        adminViewModel.adminLiveData.observe(getViewLifecycleOwner(), users -> listAdapter.setData(users));
+        adminViewModel.adminLiveData.observe(getViewLifecycleOwner(), new Observer<List<User>>() {
+            @Override
+            public void onChanged(List<User> usr) {
+                listAdapter.setData(usr);
+                // Log.d("retr",usr.get(1).getEmail());
+            }
+        });
 
     }
-
-        //return binding.getRoot();
-    //}
-
 
 
     private void initAdminViewModel() {
@@ -138,12 +139,14 @@ public class AdminFragment extends Fragment {
 
     private void Back_but(){
         Intent intent = new Intent((UsrActivity) requireActivity(), MainActivity.class);
-        SharedPreferences sharedPref = context.getSharedPreferences(SHARED_PREF,0);
+        Context context= new MainActivity();
+        SharedPreferences sharedPref = this.getActivity().getSharedPreferences(SHARED_PREF,context.MODE_PRIVATE);
         SharedPreferences.Editor editor = sharedPref.edit();
         editor.putString(EMAIL, "");
         editor.putString(PASSWORD, "");
         editor.putString(LVL, "");
         editor.apply();
+
         startActivity(intent);
     }
 }
